Bug fixing and restructuring ui

+++ src/app/modules/jobs/components/job-ad-item/job-ad-item.component.ts
@@ -18,7 +18,7 @@ export class JobAdItemComponent {
   @Input() jobAd?: JobAd;
 
   delete(): void {
-    if(this.jobAd) {
+    if(this.jobAd && this.jobAd.id) {
       this.jobService.deleteJobAd(this.jobAd.id);
     }
   }
